moved specific fields in with rest of fields, removed theming
